refactor(cluster_k8s): move Minikube code to a separate file
