+++ sdcm/cluster_k8s/__init__.py
@@ -15,7 +15,6 @@
 
 import os
 import time
-import atexit
 import logging
 import contextlib
 from datetime import datetime
@@ -23,32 +22,23 @@ from copy import deepcopy
 from typing import Optional, Union, List, Dict, Any, Literal, ContextManager
 from difflib import unified_diff
 from tempfile import NamedTemporaryFile
-from textwrap import dedent
 from functools import cached_property, partialmethod
-from itertools import chain
 from threading import RLock
 
 import yaml
 from kubernetes.dynamic.resource import Resource, ResourceField, ResourceInstance
-from invoke.exceptions import UnexpectedExit
 
-from sdcm import cluster, cluster_gce, cluster_docker
-from sdcm.cluster import ClusterNodesNotReady
-from sdcm.remote import LOCALRUNNER, shell_script_cmd
+from sdcm import cluster, cluster_docker
 from sdcm.remote.kubernetes_cmd_runner import KubernetesCmdRunner
 from sdcm.sct_config import sct_abs_path
 from sdcm.sct_events import TestFrameworkEvent
 from sdcm.utils.k8s import KubernetesOps, JSON_PATCH_TYPE
-from sdcm.utils.common import get_free_port, wait_for_port
-from sdcm.utils.decorators import log_run_info, timeout, retrying
-from sdcm.utils.docker_utils import ContainerManager
+from sdcm.utils.decorators import log_run_info, timeout
 from sdcm.utils.remote_logger import get_system_logging_thread, CertManagerLogger, ScyllaOperatorLogger
 from .operator_monitoring import ScyllaOperatorLogMonitoring, ScyllaOperatorStatusMonitoring
 from sdcm.coredump import CoredumpExportFileThread
 
 
-KUBECTL_PROXY_PORT = 8001
-KUBECTL_PROXY_CONTAINER = "auto_ssh:kubectl_proxy"
 SCYLLA_OPERATOR_CONFIG = sct_abs_path("sdcm/k8s_configs/operator.yaml")
 SCYLLA_CLUSTER_CONFIG = sct_abs_path("sdcm/k8s_configs/cluster.yaml")
 SCYLLA_API_VERSION = "scylla.scylladb.com/v1alpha1"
@@ -64,89 +54,6 @@ LOGGER = logging.getLogger(__name__)
 
 
 IptablesChainCommand = Literal["A", "C", "D"]
-
-
-class MinikubeOps:
-    @classmethod
-    def setup_minikube(cls, node: cluster.BaseNode, kubectl_version: str, minikube_version: str) -> None:
-        LOGGER.debug("Prepare %s to run Minikube cluster", node)
-        if node.distro.is_ubuntu18:
-            cls.setup_minikube_ubuntu18(node, kubectl_version, minikube_version)
-        else:
-            raise ValueError(f"{node.distro} is not supported by SCT for running Minikube")
-
-    @staticmethod
-    def setup_minikube_ubuntu18(node: cluster.BaseNode, kubectl_version: str, minikube_version: str) -> None:
-        minikube_setup_script = dedent(f"""
-            # Make sure that cloud-init finished running.
-            until [ -f /var/lib/cloud/instance/boot-finished ]; do sleep 1; done
-
-            # Disable apt-key warnings and set non-interactive frontend.
-            export APT_KEY_DONT_WARN_ON_DANGEROUS_USAGE=1
-            export DEBIAN_FRONTEND=noninteractive
-
-            apt-get -qq update
-            apt-get -qq install --no-install-recommends apt-transport-https conntrack
-
-            # Install and configure Docker.
-            curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add -
-            add-apt-repository \\"deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable\\"
-            apt-get -qq install --no-install-recommends docker-ce docker-ce-cli containerd.io
-            usermod -a -G docker {node.ssh_login_info["user"]}
-            echo fs.aio-max-nr=1048576 >> /etc/sysctl.d/99-sct-minikube.conf
-            echo net.ipv4.ip_forward=1 >> /etc/sysctl.d/99-sct-minikube.conf
-            echo net.ipv4.conf.all.forwarding=1 >> /etc/sysctl.d/99-sct-minikube.conf
-            # kernel.core_pattern is global variable, it has same value inside of docker container and on the host
-            # we patch it on the host
-            mkdir -p /var/lib/scylla/coredumps
-            echo kernel.core_pattern=/var/lib/scylla/coredumps/%h-%P-%u-%g-%s-%t.core >> /etc/sysctl.d/99-sct-minikube.conf
-            sysctl --system
-
-            # Download kubectl binary.
-            curl -fsSLo /usr/local/bin/kubectl \
-                https://storage.googleapis.com/kubernetes-release/release/v{kubectl_version}/bin/linux/amd64/kubectl
-            chmod +x /usr/local/bin/kubectl
-
-            # Download Minikube binary.
-            curl -fsSLo /usr/local/bin/minikube \
-                https://storage.googleapis.com/minikube/releases/v{minikube_version}/minikube-linux-amd64
-            chmod +x /usr/local/bin/minikube
-        """)
-        node.remoter.run(f'sudo bash -cxe "{minikube_setup_script}"', change_context=True)
-
-    @staticmethod
-    def start_minikube(node: cluster.BaseNode) -> None:
-        LOGGER.debug("Start Minikube cluster")
-        minikube_start_script = dedent(f"""
-            sysctl fs.protected_regular=0
-            minikube start --driver=none --extra-config=apiserver.service-node-port-range=1-65535
-            ip link set docker0 promisc on
-            chown -R $USER $HOME/.kube $HOME/.minikube
-        """)
-        node.remoter.run(f'sudo -E bash -cxe "{minikube_start_script}"')
-
-    @staticmethod
-    def get_kubectl_proxy(node: cluster.BaseNode) -> [str, int]:
-        LOGGER.debug("Stop any other process listening on kubectl proxy port")
-        node.remoter.run(f"sudo fuser -v4k {KUBECTL_PROXY_PORT}/tcp", ignore_status=True)
-
-        LOGGER.debug("Start kubectl proxy in detached mode")
-        node.remoter.run(
-            "setsid kubectl proxy --disable-filter --accept-hosts '.*' > proxy.log 2>&1 < /dev/null & sleep 1")
-
-        LOGGER.debug("Start auto_ssh for kubectl proxy")
-        ContainerManager.run_container(node, KUBECTL_PROXY_CONTAINER,
-                                       local_port=KUBECTL_PROXY_PORT,
-                                       remote_port=get_free_port(),
-                                       ssh_mode="-L")
-
-        host = "127.0.0.1"
-        port = int(ContainerManager.get_environ(node, KUBECTL_PROXY_CONTAINER)["SSH_TUNNEL_REMOTE"])
-
-        LOGGER.debug("Waiting for port %s:%s is accepting connections", host, port)
-        wait_for_port(host, port)
-
-        return host, port
 
 
 class KubernetesCluster:  # pylint: disable=too-few-public-methods
@@ -232,95 +139,6 @@ class KubernetesCluster:  # pylint: disable=too-few-public-methods
     def operator_pod_status(self):
         pods = KubernetesOps.list_pods(self, namespace='scylla-operator-system')
         return pods[0].status if pods else None
-
-
-class MinikubeCluster(KubernetesCluster):
-    @property
-    def minikube_version(self):
-        try:
-            return self._minikube_version
-        except AttributeError:
-            raise ValueError("You should set `minikube_version' first.") from None
-
-    @minikube_version.setter
-    def minikube_version(self, value):
-        self._minikube_version = value
-
-    @cached_property
-    def local_kubectl_version(self):  # pylint: disable=no-self-use
-        # Example of kubectl command output:
-        #   $ kubectl version --client --short
-        #   Client Version: v1.18.5
-        return LOCALRUNNER.run("kubectl version --client --short").stdout.rsplit(None, 1)[-1][1:]
-
-    @cached_property
-    def k8s_server_url(self):
-        host, port = MinikubeOps.get_kubectl_proxy(self.nodes[-1])
-        return f"http://{host}:{port}"
-
-    def node_setup(self, node, verbose=False, timeout=3600):
-        node.wait_ssh_up(verbose=verbose, timeout=timeout)
-
-        if not cluster.Setup.REUSE_CLUSTER:
-            MinikubeOps.setup_minikube(node=node,
-                                       kubectl_version=self.local_kubectl_version,
-                                       minikube_version=self.minikube_version)
-            MinikubeOps.start_minikube(node)
-
-    @cached_property
-    def hydra_dest_ip(self) -> Optional[str]:
-        return self.nodes[-1].external_address
-
-    @cached_property
-    def nodes_dest_ip(self) -> Optional[str]:
-        return self.nodes[-1].ip_address
-
-    @cached_property
-    def remoter(self) -> KubernetesCmdRunner:
-        return self.nodes[-1].remoter
-
-    def docker_pull(self, image):
-        LOGGER.info("Pull `%s' to Minikube' Docker environment", image)
-        self.remoter.run(f"docker pull -q {image}")
-
-
-class GceMinikubeCluster(MinikubeCluster, cluster_gce.GCECluster):
-    def __init__(self, minikube_version, gce_image, gce_image_type, gce_image_size, gce_network, services, credentials,  # pylint: disable=too-many-arguments
-                 gce_instance_type="n1-highmem-8", gce_image_username="centos", user_prefix=None, params=None,
-                 gce_datacenter=None):
-        # pylint: disable=too-many-locals
-        self.minikube_version = minikube_version
-
-        cluster_prefix = cluster.prepend_user_prefix(user_prefix, "k8s-minikube")
-        node_prefix = cluster.prepend_user_prefix(user_prefix, "node")
-        super().__init__(gce_image=gce_image,
-                         gce_image_type=gce_image_type,
-                         gce_image_size=gce_image_size,
-                         gce_n_local_ssd=0,
-                         gce_network=gce_network,
-                         gce_instance_type=gce_instance_type,
-                         gce_image_username=gce_image_username,
-                         services=services,
-                         credentials=credentials,
-                         cluster_prefix=cluster_prefix,
-                         node_prefix=node_prefix,
-                         n_nodes=1,
-                         add_disks=None,
-                         params=params,
-                         gce_region_names=gce_datacenter,
-                         node_type="scylla-db")
-
-    @cluster.wait_for_init_wrap
-    def wait_for_init(self):
-        for node in self.nodes:
-            node.remoter._reconnect()  # Reconnect to update user groups in main thread too.
-
-    def get_node_ips_param(self, public_ip=True):
-        raise NotImplementedError("Not implemented yet.")  # TODO: add implementation of this method
-
-    def destroy(self) -> None:
-        super().destroy()
-        self.stop_k8s_task_threads()
 
 
 class BasePodContainer(cluster.BaseNode):
@@ -513,27 +331,6 @@ class BasePodContainer(cluster.BaseNode):
         self._coredump_thread = CoredumpExportFileThread(
             self, self._maximum_number_of_cores_to_publish, ['/var/lib/scylla/coredumps'])
         self._coredump_thread.start()
-
-
-class MinicubeScyllaPodContainer(BasePodContainer):
-    parent_cluster: 'MinikubeScyllaPodCluster'
-
-    @cached_property
-    def host_remoter(self):
-        return self.parent_cluster.k8s_cluster.remoter
-
-    @property
-    def docker_id(self):
-        return self.host_remoter.run(
-            f'docker ps --filter "label=io.kubernetes.pod.name={self.name}" '
-            '--filter "label=io.kubernetes.container.name=scylla" -q').stdout.strip()
-
-    def restart(self):
-        self.host_remoter.run(f'docker restart {self.docker_id}')
-
-    @cached_property
-    def node_type(self) -> 'str':
-        return 'db'
 
 
 class PodCluster(cluster.BaseCluster):
@@ -759,102 +556,6 @@ class ScyllaPodCluster(cluster.BaseScyllaCluster, PodCluster):
     def rollout_restart(self):
         self.k8s_cluster.kubectl("rollout restart statefulset", namespace=self.namespace)
         self.wait_for_pods_readiness()
-
-
-class MinikubeScyllaPodCluster(ScyllaPodCluster):
-    k8s_cluster: MinikubeCluster
-    nodes: List[MinicubeScyllaPodContainer]
-
-    def _create_node(self, node_index: int, pod_name: str) -> MinicubeScyllaPodContainer:
-        node = MinicubeScyllaPodContainer(
-            parent_cluster=self,
-            name=pod_name,
-            base_logdir=self.logdir,
-            node_prefix=self.node_prefix,
-            node_index=node_index)
-        node.init()
-        return node
-
-    def _iptables_redirect_rules(self,
-                                 dest_ip: str,
-                                 iptables_bin: str = IPTABLES_BIN,
-                                 command: IptablesChainCommand = "A",
-                                 nodes: Optional[List[BasePodContainer]] = None) -> List[str]:
-        if nodes is None:
-            nodes = self.nodes
-        return list(chain.from_iterable(node.iptables_node_redirect_rules(dest_ip=dest_ip,
-                                                                          iptables_bin=iptables_bin,
-                                                                          command=command) for node in nodes))
-
-    def hydra_iptables_redirect_rules(self,
-                                      command: IptablesChainCommand = "A",
-                                      nodes: Optional[list] = None) -> List[str]:
-        return self._iptables_redirect_rules(dest_ip=self.k8s_cluster.hydra_dest_ip,
-                                             iptables_bin=IPTABLES_LEGACY_BIN,
-                                             command=command,
-                                             nodes=nodes)
-
-    def nodes_iptables_redirect_rules(self,
-                                      command: IptablesChainCommand = "A",
-                                      nodes: Optional[list] = None) -> List[str]:
-        return self._iptables_redirect_rules(dest_ip=self.k8s_cluster.nodes_dest_ip, command=command, nodes=nodes)
-
-    def update_nodes_iptables_redirect_rules(self,
-                                             command: IptablesChainCommand = "A",
-                                             nodes: Optional[list] = None) -> None:
-        nodes_to_update = []
-        if tester := cluster.Setup.tester_obj():
-            if tester.loaders:
-                nodes_to_update.extend(tester.loaders.nodes)
-            if tester.monitors:
-                nodes_to_update.extend(tester.monitors.nodes)
-
-        if nodes_to_update:
-            LOGGER.debug("Found following nodes to apply new iptables rules: %s", nodes_to_update)
-            iptables_rules = "\n".join(self.nodes_iptables_redirect_rules(command=command, nodes=nodes))
-            for node in nodes_to_update:
-                node.remoter.sudo(shell_script_cmd(iptables_rules))
-
-    def add_nodes(self,
-                  count: int,
-                  ec2_user_data: str = "",
-                  dc_idx: int = 0,
-                  enable_auto_bootstrap: bool = False) -> List[MinicubeScyllaPodContainer]:
-        new_nodes = super().add_nodes(count=count,
-                                      ec2_user_data=ec2_user_data,
-                                      dc_idx=dc_idx,
-                                      enable_auto_bootstrap=enable_auto_bootstrap)
-        for node in new_nodes:
-            KubernetesOps.expose_pod_ports(self.k8s_cluster, node.name,
-                                           ports=SCYLLA_POD_EXPOSED_PORTS,
-                                           labels=f"statefulset.kubernetes.io/pod-name={node.name}",
-                                           selector=f"statefulset.kubernetes.io/pod-name={node.name}",
-                                           namespace=self.namespace)
-
-        add_rules_commands = self.hydra_iptables_redirect_rules(nodes=new_nodes)
-        del_rules_commands = self.hydra_iptables_redirect_rules(command="D", nodes=new_nodes)
-
-        LOCALRUNNER.sudo(shell_script_cmd("\n".join(add_rules_commands)))
-        atexit.register(LOCALRUNNER.sudo, shell_script_cmd("\n".join(del_rules_commands)))
-
-        self.update_nodes_iptables_redirect_rules(nodes=new_nodes)
-
-        return new_nodes
-
-    def terminate_node(self, node: BasePodContainer) -> None:
-        assert self.nodes[-1] == node, "Can withdraw the last node only"
-        self.update_nodes_iptables_redirect_rules(command="D", nodes=[node, ])
-        KubernetesOps.unexpose_pod_ports(self.k8s_cluster, node.name, namespace=self.namespace)
-        super().terminate_node(node)
-
-    def upgrade_scylla_cluster(self, new_version: str) -> None:
-        self.k8s_cluster.docker_pull(f"scylladb/scylla:{new_version}")
-        return super().upgrade_scylla_cluster(new_version)
-
-    @retrying(n=20, sleep_time=60, allowed_exceptions=(ClusterNodesNotReady, UnexpectedExit),
-              message="Waiting for nodes to join the cluster")
-    def wait_for_nodes_up_and_normal(self, nodes, verification_node=None):
-        super().wait_for_nodes_up_and_normal(nodes, verification_node)
 
 
 def iptables_port_redirect_rule(iptables_bin: str,
